Split up Pokemon data types to own files
